[Progress] read / write, fixed some bugs on offset read / write

+++ sfs_client/src/client/client_syscall.rs
@@ -323,7 +323,7 @@ fn internal_pread(f: Arc<Mutex<OpenFile>>, buf: * mut c_char, count: i64, offset
         error_msg("client::sfs_pread".to_string(), format!("read with error {}", read_res.0));
         return -1;
     }
-    return read_res.1;
+    return read_res.1 as i64;
 }
 
 #[no_mangle]
